Move line parser code together
